AccountingFirm/requirements.txt
Flask==2.0.1
Pillow>=10.0.0
Werkzeug==2.0.1
python-dotenv==0.19.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
google-cloud-vision>=3.4.4
flask-cors==3.0.10 